Changed note system to be more flexible and to support read-only
Changed the note system so that almost all responsibility on drawing
notes is delegated to the note-handler. Only adding new notes and
delegating events to clear, load or save-state, and on mode-changes
are handled by the general annotator. Actual implementation is done
by note-handler. This is to allow flexible showing of the notes and
to for example more easy allow saving files to some underlying
database etc. that the annotator does not need to know about.